+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `BitSetN::is`
   - `BitSetN::is_not`
   - `BitSetN::is_disjoint`
+  - `BitSetN::is_subset`
   - `<BitSetN as Debug>::fmt`
 - Eight shift methods for `BitSetN`
   - `shifted_up_by`
@@ -31,6 +32,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `<BitSetN as LowerHex>::fmt`
   - `<BitSetN as Octal>::fmt`
   - `<BitSetN as UpperHex>::fmt`
+
+### Changed
+
+- Some bits of the test bitsets
 
 ## [0.1.0-beta] - 2025-04-11